+++ src/logic/Logic/Procedures/GenerateCode.hs
@@ -4,7 +4,6 @@ import AlgebraicPath qualified as Path
 import Control.Monad.Parallel qualified as MonadParallel
 import Data.Aeson.Text qualified as Aeson.Text
 import Data.Map.Strict qualified as Map
-import Data.Text qualified as Text
 import Logic.Capabilities.Fs (FsOps (..))
 import Logic.Capabilities.GeneratorRuntime (LoadsGen (..))
 import Logic.Capabilities.Reporting (Warns (..))
@@ -16,7 +15,6 @@ import Logic.Domain.Report (Report (..))
 import PGenieGen qualified as Gen
 import PGenieGen.Model.Input qualified as Gen.Input
 import PGenieGen.Model.Output qualified as Gen.Output
-import PGenieGen.Model.Output.Report qualified as Gen.Output.Report
 import Utils.Prelude hiding (readFile, writeFile)
 
 type Port m = (MonadParallel m, Stages m, Warns m, FsOps m, MonadError Report m, LoadsGen m)
@@ -71,21 +69,16 @@ run Params {projectFile, project} =
           stage "Compiling" 0 do
             let (compileFn, genUrl, newHash) = compileFnWithHash
                 output = compileFn project
-            case output.result of
-              Gen.Output.ErrResult report ->
+            case output of
+              Gen.Output.ErrOutput report ->
                 throwError
                   ( Report
                       report.path
                       report.message
                       Nothing
-                      [ ( "warnings",
-                          output.warnings
-                            & map Gen.Output.Report.toWarningYamlText
-                            & Text.intercalate "\n"
-                        )
-                      ]
+                      []
                   )
-              Gen.Output.OkResult generatedFiles -> do
+              Gen.Output.OkOutput Gen.Output.OutputOk {warnings, value = generatedFiles} -> do
                 artifactPath <- case Path.maybeFromText name of
                   Nothing ->
                     throwError
@@ -101,7 +94,7 @@ run Params {projectFile, project} =
                   let modifiedPath = artifactPath <> file.path
                   writeFile modifiedPath file.content
                   pure modifiedPath
-                pure (Artifact name output.warnings generatedFilePaths, (genUrl, newHash))
+                pure (Artifact name warnings generatedFilePaths, (genUrl, newHash))
 
     let (artifacts, hashPairs) = unzip artifactsWithHashes
         updatedHashes = Map.union (Map.fromList hashPairs) existingHashes